pyproject.toml
[tool.poetry]
name = "tool"
version = "0.1.0"
description = ""
authors = ["makito <imakitos1023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.3.0"
ruff = "^0.0.267"
pytest = "^7.3.1"
beautifulsoup4 = "^4.12.2"
googlemaps = "^4.10.0"
python-dotenv = "^1.0.0"
requests = "2.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
